Fix sad-sycl and seam-carving-sycl - add image library stubs
sad-sycl:
- Created minimal bitmap_image.hpp stub for BMP file loading
- Added #include <cmath> and changed ceilf to std::ceil
- Provides basic bitmap_image class with width(), height(), get_pixel()
- Full library available at https://github.com/ArashPartow/bitmap

seam-carving-sycl:
- Created minimal stb_image.h stub for image loading
- Created minimal stb_image_write.h stub for BMP writing
- Implements stbi_load() and stbi_write_bmp() functions
- Full library available at https://github.com/nothings/stb

Both implementations are simplified stubs that allow compilation
and basic testing. For production use, the full external libraries
should be used.

+++ src/sad-sycl/main.cpp
@@ -1,4 +1,5 @@
 #include <iostream>
+#include <cmath>
 #include <chrono>
 #include <sycl/sycl.hpp>
 #include "bitmap_image.hpp"
@@ -208,11 +209,11 @@ int main(int argc, char* argv[]) {
   int *d_min_mse = sycl::malloc_device<int>(1, q);
   int *d_num_occurances = sycl::malloc_device<int>(1, q);
 
-  sycl::range<2> gws ((unsigned int)ceilf((float)main_height / BLOCK_SIZE_Y) * BLOCK_SIZE_Y,
-                      (unsigned int)ceilf((float)main_width / BLOCK_SIZE_X) * BLOCK_SIZE_X );
+  sycl::range<2> gws ((unsigned int)std::ceil((float)main_height / BLOCK_SIZE_Y) * BLOCK_SIZE_Y,
+                      (unsigned int)std::ceil((float)main_width / BLOCK_SIZE_X) * BLOCK_SIZE_X );
   sycl::range<2> lws (BLOCK_SIZE_Y, BLOCK_SIZE_X);
 
-  sycl::range<1> gws2 ((unsigned int)ceilf((float)sad_array_size / BLOCK_SIZE) * BLOCK_SIZE);
+  sycl::range<1> gws2 ((unsigned int)std::ceil((float)sad_array_size / BLOCK_SIZE) * BLOCK_SIZE);
   sycl::range<1> lws2 (BLOCK_SIZE);
 
   // Measure device execution time
